+++ src/StoredEvents/EventSubscriber.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace Spatie\EventSourcing;
+namespace Spatie\EventSourcing\StoredEvents;
+
+use Spatie\EventSourcing\StoredEvents\Repositories\StoredEventRepository;
 
 class EventSubscriber
 {
@@ -22,6 +24,10 @@ class EventSubscriber
             return;
         }
 
+        if ($this->isFiredFromAggregateRoot($payload[0])) {
+            return;
+        }
+
         $this->storeEvent($payload[0]);
     }
 
@@ -38,5 +44,10 @@ class EventSubscriber
         }
 
         return is_subclass_of($event, ShouldBeStored::class);
+    }
+
+    private function isFiredFromAggregateRoot($event): bool
+    {
+        return $event->firedFromAggregateRoot ?? false;
     }
 }