+++ src/StoredEvents/Models/EloquentStoredEvent.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Spatie\EventSourcing\Models;
+namespace Spatie\EventSourcing\StoredEvents\Models;
 
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
-use Spatie\EventSourcing\ShouldBeStored;
-use Spatie\EventSourcing\StoredEvent;
+use Spatie\EventSourcing\StoredEvents\ShouldBeStored;
+use Spatie\EventSourcing\StoredEvents\StoredEvent;
 use Spatie\SchemalessAttributes\SchemalessAttributes;
 
 class EloquentStoredEvent extends Model
@@ -20,6 +20,8 @@ class EloquentStoredEvent extends Model
         'event_properties' => 'array',
         'meta_data' => 'array',
     ];
+    
+    protected ?ShouldBeStored $originalEvent = null;
 
     public function toStoredEvent(): StoredEvent
     {
@@ -27,15 +29,23 @@ class EloquentStoredEvent extends Model
             'id' => $this->id,
             'event_properties' => $this->event_properties,
             'aggregate_uuid' => $this->aggregate_uuid ?? '',
+            'aggregate_version' => $this->aggregate_version ?? 0,
             'event_class' => $this->event_class,
             'meta_data' => $this->meta_data,
             'created_at' => $this->created_at,
-        ]);
+        ], $this->originalEvent);
+    }
+    
+    public function setOriginalEvent(ShouldBeStored $event): self
+    {
+        $this->originalEvent = $event;
+        
+        return $this;
     }
 
-    public function getEventAttribute(): ShouldBeStored
+    public function getEventAttribute(): ?ShouldBeStored
     {
-        return $this->toStoredEvent()->event;
+        return $this->originalEvent ??= $this->toStoredEvent()->event;
     }
 
     public function getMetaDataAttribute(): SchemalessAttributes