+++ src/StoredEvents/StoredEvent.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Spatie\EventSourcing;
+namespace Spatie\EventSourcing\StoredEvents;
 
 use Exception;
 use Illuminate\Contracts\Support\Arrayable;
@@ -8,7 +8,6 @@ use Illuminate\Support\Arr;
 use Spatie\EventSourcing\EventSerializers\EventSerializer;
 use Spatie\EventSourcing\Exceptions\InvalidStoredEvent;
 use Spatie\EventSourcing\Facades\Projectionist;
-use Spatie\SchemalessAttributes\SchemalessAttributes;
 
 class StoredEvent implements Arrayable
 {
@@ -19,33 +18,28 @@ class StoredEvent implements Arrayable
 
     public string $aggregate_uuid;
 
+    public string $aggregate_version;
+
     public string $event_class;
 
-    public SchemalessAttributes $meta_data;
+    /** @var array|string */
+    public $meta_data;
 
     public string $created_at;
 
     public ?ShouldBeStored $event;
 
-    public function __construct(array $data)
+    public function __construct(array $data, ?ShouldBeStored $originalEvent = null)
     {
         $this->id = $data['id'] ?? null;
         $this->event_properties = $data['event_properties'];
         $this->aggregate_uuid = $data['aggregate_uuid'];
+        $this->aggregate_version = $data['aggregate_version'];
         $this->event_class = self::getActualClassForEvent($data['event_class']);
         $this->meta_data = $data['meta_data'];
         $this->created_at = $data['created_at'];
-
-        try {
-            $this->event = app(EventSerializer::class)->deserialize(
-                self::getActualClassForEvent($this->event_class),
-                is_string($this->event_properties)
-                    ? $this->event_properties
-                    : json_encode($this->event_properties)
-            );
-        } catch (Exception $exception) {
-            throw InvalidStoredEvent::couldNotUnserializeEvent($this, $exception);
-        }
+        
+        $this->instantiateEvent($originalEvent);
     }
 
     public function toArray()
@@ -54,15 +48,28 @@ class StoredEvent implements Arrayable
             'id' => $this->id,
             'event_properties' => $this->event_properties,
             'aggregate_uuid' => $this->aggregate_uuid,
+            'aggregate_version' => $this->aggregate_version,
             'event_class' => self::getEventClass($this->event_class),
             'meta_data' => $this->meta_data instanceof Arrayable ? $this->meta_data->toArray() : (array) $this->meta_data,
             'created_at' => $this->created_at,
         ];
     }
 
+    public function handleForAggregateRoot(): void
+    {
+        $this->handle();
+
+        if (! config('event-sourcing.dispatch_events_from_aggregate_roots', false)) {
+            return;
+        }
+
+        $this->event->firedFromAggregateRoot = true;
+        event($this->event);
+    }
+
     public function handle()
     {
-        Projectionist::handleWithSyncProjectors($this);
+        Projectionist::handleWithSyncEventHandlers($this);
 
         if (method_exists($this->event, 'tags')) {
             $tags = $this->event->tags();
@@ -78,20 +85,46 @@ class StoredEvent implements Arrayable
             $tags ?? []
         );
 
-        dispatch($storedEventJob->onQueue($this->event->queue ?? config('event-sourcing.queue')));
+        dispatch($storedEventJob->onQueue($this->getQueueName()));
+    }
+
+    /** @psalm-suppress TypeDoesNotContainType */
+    protected function getQueueName(): ?string
+    {
+        return $this->event->queue ?? config('event-sourcing.queue');
     }
 
     protected function shouldDispatchJob(): bool
     {
-        if (Projectionist::getAsyncProjectorsFor($this)->isNotEmpty()) {
-            return true;
-        }
+        /** @var \Spatie\EventSourcing\EventHandlers\EventHandlerCollection $eventHandlers */
+        $eventHandlers = Projectionist::allEventHandlers();
 
-        if (Projectionist::getReactorsFor($this)->isNotEmpty()) {
-            return true;
-        }
+        return $eventHandlers->asyncEventHandlers()->count() > 0;
+    }
+    
+    protected function instantiateEvent(?ShouldBeStored $originalEvent): void
+    {
+        if ($originalEvent) {
+            $this->event = $originalEvent;
 
-        return false;
+            return;
+        }
+    
+        try {
+            $this->event = app(EventSerializer::class)->deserialize(
+                self::getActualClassForEvent($this->event_class),
+                is_string($this->event_properties)
+                    ? $this->event_properties
+                    : json_encode($this->event_properties),
+                is_string($this->meta_data)
+                    ? $this->meta_data
+                    : json_encode($this->meta_data),
+            );
+        
+            $this->event->setMetaData(optional($this->meta_data)->toArray());
+        } catch (Exception $exception) {
+            throw InvalidStoredEvent::couldNotUnserializeEvent($this, $exception);
+        }
     }
 
     protected static function getActualClassForEvent(string $class): string