+++ tests/TestClasses/AggregateRoots/DummyAggregateRoot.php
@@ -2,7 +2,7 @@
 
 namespace Spatie\EventSourcing\Tests\TestClasses\AggregateRoots;
 
-use Spatie\EventSourcing\AggregateRoot;
+use Spatie\EventSourcing\AggregateRoots\AggregateRoot;
 use Spatie\EventSourcing\Tests\TestClasses\AggregateRoots\StorableEvents\DummyEvent;
 
 class DummyAggregateRoot extends AggregateRoot
@@ -14,8 +14,13 @@ class DummyAggregateRoot extends AggregateRoot
         $this->recordThat(new DummyEvent($this->latestInteger + 1));
     }
 
-    public function applyDummyEvent(DummyEvent $dummyEvent)
+    protected function applyDummyEvent(DummyEvent $dummyEvent)
     {
         $this->latestInteger = $dummyEvent->integer;
+    }
+
+    public function getLatestInteger(): int
+    {
+        return $this->latestInteger;
     }
 }