+++ CHANGELOG.md
@@ -2,10 +2,104 @@
 
 All notable changes to `laravel-event-sourcing` will be documented in this file:
 
+## 4.6.0 - 2020-10-21
+
+- allow protected apply methods (#136)
+
+## 4.5.3 - 2020-10-02
+
+- re-use existing instance of `ShouldBeStored` when possible (#158)
+
+## 4.5.2 - 2020-10-02
+
+- fix Paths and Reference URL in event-sourcing.php config file (#159)
+
+## 4.5.1 - 2020-09-27
+
+- added `loadUuid` (#156)
+
+## 4.5.0 - 2020-09-25
+
+- make normalizers configurable (#153)
+
+## 4.4.0 - 2020-09-18
+
+- Support `then` for aggregate root fakes (#154)
+
+## 4.3.1 - 2020-09-09
+
+- Support Laravel 8
+
+## 4.3.0 - 2020-08-24
+
+- support Carbon dates in events (#137)
+
+## 4.2.0 - 2020-08-06
+
+- allow events to be dispatched from an aggregate root (#135)
+
+## 4.1.0 - 2020-08-03
+
+- add assertion that specific event is recorded (#134)
+
+## 4.0.2 - 2020-07-29
+
+- config style fix
+
+## 4.0.1 - 2020-07-29
+
+- add `snapshot_model` config key
+
+## 4.0.0 - 2020-07-22
+
+- projectors now are abstract classes instead of interfaces
+- reactors can now be easily defined by extending the reactor base class
+- projectors and reactors can be marked as async by implementing the `ShouldQueue` marker interface
+- events that extend `ShouldBeStored` now can retrieve the aggregate root uuid using `aggregateRootUuid()`
+- the package has been restructured. Namespaces of most classes have been updated.
+- events that extend `ShouldBeStored` can now handle metadata using `metaData` and `setMetaData`
+- aggregate roots can now be persisted without calling event handlers using `persistWithoutApplyingToEventHandlers`
+- the projectionist can now handle manually specified events using `handleStoredEvents`
+- added `persistAggregateRootsInTransaction` to `AggregateRoot`
+- you can now get the `uuid` of an aggregate root using the `uuid()` method
+- the `reset` method has been removed on projectors
+- the `fake` method on an aggregate root now accepts a uuid instead of an array of events
+- the `meta_data` property on `StoredEvent` is now an array or a string instead of `SchemalessAttributes`
+- apply methods on aggregates can now make use of method injection
+- pass metadata to serializer to allow events to be upgraded (#112)
+
+## 3.2.3 - 2020-07-14
+
+- default to `BigIncrements` on package table stubs (#124)
+
+## 3.2.2  - 2020-07-14
+
+- replace model where clause with uuid model scope (#123)
+
+## 3.2.1 - 2020-07-09
+
+- config file comment corrections (#121)
+
+## 3.2.0 - 2020-06-30
+
+- expose `aggregate_version` of `StoredEvent` (#115)
+
+## 3.1.8 - 2020-06-28
+
+- use `app` helper (#117)
+>>>>>>> master
+
+## 3.1.7 - 2020-06-18
+
+- allow aggregate roots to have dependencies in constructor (#111)
+
+## 3.1.6 - 2020-06-17
+
+- wrong tag, nothing changed
+
 ## 3.1.5 - 2020-05-23
 
 - only include public properties of the aggregate when snapshotting (#105)
-
 
 ## 3.1.4 - 2020-05-07
 
@@ -29,7 +123,7 @@ All notable changes to `laravel-event-sourcing` will be documented in this file:
 
 ## 3.0.4 - 2020-02-23
 
-- fix for serializing events that use immutable datatime objects (#67)
+- fix for serializing events that use immutable datetime objects (#67)
 
 ## 3.0.3 - 2020-02-18